fix: keep watermark conversion hook internal
