+++ packages/core/src/prosemirror/conversion/watermarkHost.ts
@@ -1,10 +1,6 @@
-import { panic } from "better-result";
-import type { Node as PMNode } from "prosemirror-model";
-
 import type { BlockContent } from "../../types/document";
-import { headerFooterToProseDocInternal, type ToProseDocOptions } from "./toProseDoc";
-
-const DETACHED_WATERMARK_HOST_ATTR = "_detachedWatermarkHost";
+import { headerFooterToProseDoc, type ToProseDocOptions } from "./toProseDoc";
+import { markDetachedWatermarkHost } from "./watermarkHostMarker";
 
 type WatermarkHostConversionOptions = {
   content: BlockContent[];
@@ -17,29 +13,9 @@ export const headerFooterToProseDocWithDetachedWatermarkHost = ({
   content,
   hostBlockIndex,
   options,
-}: WatermarkHostConversionOptions) =>
-  headerFooterToProseDocInternal({
-    content,
-    options,
-    detachedWatermarkHostBlockIndex: hostBlockIndex,
-  });
-
-/** Read the internal host marker without exposing it through public paragraph attrs. */
-export const expectDetachedWatermarkHostAttr = (node: PMNode): boolean => {
-  if (node.type.name !== "paragraph") {
-    panic(
-      `Invalid ProseMirror detached watermark host attrs:\nnode.type.name: Expected paragraph, got ${node.type.name}.`,
-    );
-  }
-
-  const value = Reflect.get(node.attrs, DETACHED_WATERMARK_HOST_ATTR);
-  if (value === null || value === undefined) {
-    return false;
-  }
-  if (typeof value !== "boolean") {
-    panic(
-      "Invalid ProseMirror detached watermark host attrs:\nparagraph.attrs._detachedWatermarkHost: Expected a boolean.",
-    );
-  }
-  return value;
+}: WatermarkHostConversionOptions) => {
+  const markedContent = content.map((block, blockIndex) =>
+    blockIndex === hostBlockIndex ? markDetachedWatermarkHost(block) : block,
+  );
+  return headerFooterToProseDoc(markedContent, options);
 };
